+++ packages/astro-template/astro.config.ts
@@ -4,7 +4,4 @@ import react from '@astrojs/react';
 // https://astro.build/config
 export default defineConfig({
   integrations: [react()],
-  experimental: {
-    devOverlay: true,
-  },
 });